+++ src/main/java/KGUcapstone/OutDecision/domain/user/auth/dto/RefreshToken.java
@@ -1,0 +1,29 @@
+package KGUcapstone.OutDecision.domain.user.auth.dto;
+
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.data.redis.core.RedisHash;
+import org.springframework.data.redis.core.index.Indexed;
+
+import java.io.Serializable;
+
+@Getter
+@AllArgsConstructor
+@RedisHash(value = "jwtToken", timeToLive = 60 * 60 * 24 * 14)
+public class RefreshToken implements Serializable {
+
+    @Id
+    private String id;
+
+    // findByAccessToken 이용 가능
+    @Indexed
+    private String accessToken;
+
+    private String refreshToken;
+
+    public void updateAccessToken(String accessToken) {
+        this.accessToken = accessToken;
+    }
+
+}
