+++ src/main/java/KGUcapstone/OutDecision/domain/user/auth/dto/RegisterRequestDto.java
@@ -1,0 +1,13 @@
+package KGUcapstone.OutDecision.domain.user.auth.dto;
+
+import lombok.*;
+
+@Getter
+public class RegisterRequestDto {
+    private String name;
+    private String nickname;
+    private String email;
+    private String password;
+    private String phone;
+    private String userImg;
+}
