--- src/main/java/KGUcapstone/OutDecision/domain/user/security/config/ProjectBeanConfig.java
@@ -1,38 +1,0 @@
-package KGUcapstone.OutDecision.domain.user.security.config;
-
-import KGUcapstone.OutDecision.domain.user.repository.MemberRepository;
-import KGUcapstone.OutDecision.domain.user.service.FindMemberService;
-import KGUcapstone.OutDecision.domain.user.security.service.CustomUserDetailsService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-@RequiredArgsConstructor
-public class ProjectBeanConfig {
-
-    private final MemberRepository memberRepository;
-    private final FindMemberService findMemberService;
-
-    @Bean
-    public AuthenticationProvider authenticationProvider(){
-        DaoAuthenticationProvider provider = new DaoAuthenticationProvider();
-        provider.setPasswordEncoder(passwordEncoder());
-        provider.setUserDetailsService(userService());
-        return provider;
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder(){
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public CustomUserDetailsService userService() {
-        return new CustomUserDetailsService(findMemberService, passwordEncoder(), memberRepository);
-    }
-}
