--- src/main/java/KGUcapstone/OutDecision/domain/user/security/dto/CustomUserDetails.java
@@ -1,52 +1,0 @@
-package KGUcapstone.OutDecision.domain.user.security.dto;
-
-import KGUcapstone.OutDecision.domain.user.domain.Member;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-import java.util.List;
-
-@RequiredArgsConstructor
-@Getter
-public class CustomUserDetails implements UserDetails {
-
-    private final Member member;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of(()->"ROLE_USER");
-    }
-
-    @Override
-    public String getPassword() {
-        return member.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return member.getEmail();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
