--- src/main/java/KGUcapstone/OutDecision/domain/user/security/dto/GeneratedToken.java
@@ -1,14 +1,0 @@
-package KGUcapstone.OutDecision.domain.user.security.dto;
-
-import lombok.*;
-
-@NoArgsConstructor
-@Getter
-@ToString
-@AllArgsConstructor
-@Builder
-public class GeneratedToken {
-
-    private String accessToken;
-    private String refreshToken;
-}
