--- src/main/java/KGUcapstone/OutDecision/domain/user/security/filter/JwtExceptionFilter.java
@@ -1,34 +1,0 @@
-package KGUcapstone.OutDecision.domain.user.security.filter;
-
-import KGUcapstone.OutDecision.global.error.exception.ApiResponse;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
-import org.springframework.security.oauth2.jwt.JwtException;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-@RequiredArgsConstructor
-@Component
-public class JwtExceptionFilter extends OncePerRequestFilter {
-
-    private final ObjectMapper objectMapper;
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        try {
-            filterChain.doFilter(request, response);
-        } catch (JwtException e) {
-            response.setStatus(401);
-            response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-            response.setCharacterEncoding("UTF-8");
-            objectMapper.writeValue(response.getWriter(), ApiResponse.onFailure("401",e.getMessage(),null));
-        }
-    }
-}
