--- src/main/java/KGUcapstone/OutDecision/domain/user/security/properties/JwtProperties.java
@@ -1,12 +1,0 @@
-package KGUcapstone.OutDecision.domain.user.security.properties;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Data
-@Component
-@ConfigurationProperties(prefix = "spring.jwt")
-public class JwtProperties {
-    private String secret;
-}
