+++ src/main/java/KGUcapstone/OutDecision/global/common/properties/JwtProperties.java
@@ -1,0 +1,12 @@
+package KGUcapstone.OutDecision.global.common.properties;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Data
+@Component
+@ConfigurationProperties(prefix = "spring.jwt")
+public class JwtProperties {
+    private String secret;
+}
