+++ src/main/java/KGUcapstone/OutDecision/global/security/dto/GeneratedToken.java
@@ -1,0 +1,14 @@
+package KGUcapstone.OutDecision.global.security.dto;
+
+import lombok.*;
+
+@NoArgsConstructor
+@Getter
+@ToString
+@AllArgsConstructor
+@Builder
+public class GeneratedToken {
+
+    private String accessToken;
+    private String refreshToken;
+}
