+++ src/main/java/KGUcapstone/OutDecision/global/security/dto/SecurityUserDto.java
@@ -1,0 +1,15 @@
+package KGUcapstone.OutDecision.global.security.dto;
+
+import lombok.*;
+
+@NoArgsConstructor
+@Getter
+@ToString
+@AllArgsConstructor
+@Builder
+public class SecurityUserDto {
+    private String email;
+    private String nickname;
+    private String role;
+    private Long memberId;
+}
