+++ src/main/java/KGUcapstone/OutDecision/global/security/filter/JwtAuthFilter.java
@@ -1,0 +1,80 @@
+package KGUcapstone.OutDecision.global.security.filter;
+
+import KGUcapstone.OutDecision.domain.user.domain.Member;
+import KGUcapstone.OutDecision.global.security.dto.SecurityUserDto;
+import KGUcapstone.OutDecision.domain.user.repository.MemberRepository;
+import KGUcapstone.OutDecision.global.common.util.JwtUtil;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.jwt.JwtException;
+import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import java.io.IOException;
+import java.util.List;
+
+@RequiredArgsConstructor
+@Slf4j
+@Component
+public class JwtAuthFilter extends OncePerRequestFilter {
+
+    private final JwtUtil jwtUtil;
+    private final MemberRepository memberRepository;
+
+    @Override
+    protected boolean shouldNotFilter(HttpServletRequest request) throws ServletException{
+        return request.getRequestURI().contains("/token/");
+    }
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+        // request Header에서 AccessToken을 가져온다.
+        String atc = request.getHeader("Authorization");
+
+        // 토큰 검사 생략(모두 허용 URL의 경우 토큰 검사 통과)
+        if (!StringUtils.hasText(atc)) {
+            doFilter(request, response, filterChain);
+            return;
+        }
+
+        // AccessToken을 검증하고, 만료되었을경우 예외를 발생시킨다.
+        if (!jwtUtil.verifyToken(atc)) {
+            throw new JwtException("Access Token 만료!");
+        }
+
+        // AccessToken의 값이 있고, 유효한 경우에 진행한다.
+        if (jwtUtil.verifyToken(atc)) {
+
+            // AccessToken 내부의 payload에 있는 email로 user를 조회한다.
+            Member findMember = memberRepository.findByEmail(jwtUtil.getUid(atc));
+
+            // SecurityContext에 등록할 User 객체를 만들어준다.
+            SecurityUserDto userDto = SecurityUserDto.builder()
+                    .memberId(findMember.getId())
+                    .email(findMember.getEmail())
+                    .role("ROLE_".concat(findMember.getUserRole()))
+                    .nickname(findMember.getNickname())
+                    .build();
+
+            // SecurityContext에 인증 객체를 등록해준다.
+            Authentication auth = getAuthentication(userDto);
+            SecurityContextHolder.getContext().setAuthentication(auth);
+        }
+
+        filterChain.doFilter(request, response);
+    }
+
+    public Authentication getAuthentication(SecurityUserDto member) {
+        return new UsernamePasswordAuthenticationToken(member, "",
+                List.of(new SimpleGrantedAuthority(member.getRole())));
+    }
+}
