+++ src/main/java/KGUcapstone/OutDecision/global/security/handler/CustomAuthenticationFailureHandler.java
@@ -1,0 +1,21 @@
+package KGUcapstone.OutDecision.global.security.handler;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.AuthenticationFailureHandler;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+@Component
+@Slf4j
+public class CustomAuthenticationFailureHandler implements AuthenticationFailureHandler {
+    @Override
+    public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException {
+        // 인증 실패시 메인 페이지로 이동
+        response.sendRedirect("http://localhost:8080/");
+    }
+
+}
