--- src/main/java/KGUcapstone/OutDecision/global/util/CookieUtil.java
@@ -1,23 +1,0 @@
-package KGUcapstone.OutDecision.global.util;
-
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseCookie;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class CookieUtil {
-
-    public static void addCookie(HttpServletResponse response, String name, String value, int maxAge) {
-        ResponseCookie cookie = ResponseCookie.from(name, value)
-                .path("/")
-                .sameSite("None")
-                .httpOnly(true)
-                .secure(true)
-                .maxAge(maxAge)
-                .build();
-
-        response.addHeader("Set-Cookie", cookie.toString());
-    }
-}
